+++ spec/spec_helper.rb
@@ -43,4 +43,5 @@ RSpec.configure do |config|
 end
 
 require 'awesome_print'
+require 'logger'
 require 'pd' # tests both `require 'pd'` and `require 'puts_debuggerer'`